requirements.txt
allure-behave~=2.8.31
argparse
behave~=1.2.6
configcat-client>=5.0.0
configparser~=5.0.0
contextvars
cx-Oracle
Django
mysql-connector-python~=8.0.22
mysql_connector
pactman>=2.28.0
pdfminer
pip
PyPDF2
python-dateutil~=2.8.1
pytz
requests~=2.25.1
selenium~=3.141.0
setuptools
urllib3
